logic-worker/requirements.txt
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
kafka-python==2.0.2
openai==1.55.3
httpx==0.27.2
boto3==1.37.38